Jsem zvědavý na lintr kolik napíše procent u toho přihlášení
